removed redundant id's for tabs. made tabs available from a gui container with a getter. replaced all occurences of the field for tabs in survivaltablist with getter. refactored some names to have more accurate names. the guicontainer's list of available tabs is now a copy of the registered tabs so it has no consequence adding or modifying it when a gui is opened. (paired with removed tab id's this could allow adding tabs when having items in the inventory, or the like)
renamed SurvivalTabs and SurvivalTabList to more appropriate names;
renamed survivaltabs field in GuiContainer to guiTabs for more appropriate naming.
made that field protected.
